# Use serde_bytes to optimize (de)serilization speed of 8-bit filters