--- a/Cargo.toml
+++ b/Cargo.toml
@@ -30,6 +30,7 @@ repository = "ayazhafiz/xorf"
 [dependencies]
 libm = { version = "0.2.1", optional = true }
 serde = { version = "1.0.104", optional = true, features = ["derive"] }
+serde_bytes = { version = "0.11.12", optional = true }
 bincode = { version = "2.0.0-rc.3", default-features = false, optional = true, features = ["std", "derive"]} 
 num-traits = { version = "0.2.12", optional = true }
 rand = { version = "0.8", optional = true }
@@ -79,3 +80,4 @@ harness = false
 default = ["uniform-random", "binary-fuse"]
 uniform-random = ["rand"]
 binary-fuse = ["libm"]
+serde = ["dep:serde", "serde_bytes"]