--- a/Cargo.toml
+++ b/Cargo.toml
@@ -39,6 +39,7 @@ rand = { version = "0.8", optional = true }
 criterion = "0.3.0"
 criterion-macro = "0.3.0"
 rand = "0.8"
+bincode = { version = "2.0.0-rc.3", default-features = false, features = ["std", "derive", "serde"]}
 
 [[bench]]
 name = "bfuse32"
@@ -51,6 +52,7 @@ harness = false
 [[bench]]
 name = "bfuse8"
 harness = false
+required-features = ["serde"]
 
 [[bench]]
 name = "fuse32"