```
BinaryFuse8/serialize/500000
                        time:   [263.42 µs 269.47 µs 272.86 µs]
BinaryFuse8/deserialize/500000
                        time:   [1.2453 ms 1.2498 ms 1.2553 ms]
```

```
BinaryFuse8/serialize/500000
                        time:   [9.4238 µs 9.4404 µs 9.4479 µs]
                        change: [-96.524% -96.466% -96.401%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 10 measurements (20.00%)
  2 (20.00%) low mild
BinaryFuse8/deserialize/500000
                        time:   [14.670 µs 14.727 µs 14.801 µs]
                        change: [-98.822% -98.816% -98.809%] (p = 0.00 < 0.05)
                        Performance has improved
```